python3-cassandra-driver: add geomet to RDEPENDS, remove DISTUTILS options
python3-cassandra-driver is failing on do_install because the recipe is
trying to pull sources while installing, but it can no longer do this
because a recent change to python3-setuptools removing support for the
deprecated ssl.match_hostname method. Since this only seems to occur
when the DISTUTILS = "--no-libev" option to disable the libev extension
is given, remove that from the recipe and let the extension be compiled
into the package. Additionally, add python3-geomet (the dependency that
cassandra-driver was trying to pull during do_install) to the RDEPENDS
now that it exists as a recipe.

Signed-off-by: Trevor Gamblin <trevor.gamblin@windriver.com>
Signed-off-by: Khem Raj <raj.khem@gmail.com>